// src/types/BotSetting.ts
// Generated by https://quicktype.io

export default interface BotSetting {
    count: number | null;
    next: number | null;
    previous: number | null;
    results: BotSettingResult[];
}

export interface BotSettingResult {
    // Custom FIELD
    checkbox?: string;
    profitAndTotalProfit?: string;
    profitLoss?: string;
    volume_per_day?: string;
    master_above?: string;
    actions?: string;

    // FIELDs From API
    id: number;
    master: number;
    master_name: string;
    account_type: string;
    follower: number;
    follower_name: string;
    follower_username: string;
    o_amount: number;
    o_owner: number;
    fold_command: number;
    status: string;
    block_status: string;
    created_at: string;
    updated_at: string;
    current_profit: number;
    current_date_profit: number;
    current_volume: number;
    current_date_volume: number;
    aim_min: number;
    aim_max: number;
}