// src/types/responses/DashboardResponse.ts
// Generated by https://quicktype.io

export default interface DashboardResponse {
    id: number;
    date: string;
    owner: number;
    current_profit: number;
    volume: number;
    total_win: number;
    total_lose: number;
    account_type: string;
    profit: number;
    total_draft: number;
    created_at: string;
    updated_at: string;
}